refactor: introduced remove collaborator modal
